+++ packages/web/lib/index.js
@@ -48,7 +48,13 @@ function staller(duration) {
         setTimeout(resolve, duration);
     });
 }
-function fetchData(connection, body, processFunc) {
+// This API is still a work in progress; the future changes will likely be:
+// - ConnectionInfo => FetchDataRequest<T = any>
+// - FetchDataRequest.body? = string | Uint8Array | { contentType: string, data: string | Uint8Array }
+//   - If string => text/plain, Uint8Array => application/octet-stream (if content-type unspecified)
+// - FetchDataRequest.processFunc = (body: Uint8Array, response: FetchDataResponse) => T
+// For this reason, it should be considered internal until the API is finalized
+function _fetchData(connection, body, processFunc) {
     // How many times to retry in the event of a throttle
     var attemptLimit = (typeof (connection) === "object" && connection.throttleLimit != null) ? connection.throttleLimit : 12;
     logger.assertArgument((attemptLimit > 0 && (attemptLimit % 1) === 0), "invalid connection throttle limit", "connection.throttleLimit", attemptLimit);
@@ -259,7 +265,7 @@ function fetchData(connection, body, processFunc) {
     })();
     return Promise.race([runningTimeout.promise, runningFetch]);
 }
-exports.fetchData = fetchData;
+exports._fetchData = _fetchData;
 function fetchJson(connection, json, processFunc) {
     var processJsonFunc = function (value, response) {
         var result = null;
@@ -299,7 +305,7 @@ function fetchJson(connection, json, processFunc) {
         }
         connection = updated;
     }
-    return fetchData(connection, body, processJsonFunc);
+    return _fetchData(connection, body, processJsonFunc);
 }
 exports.fetchJson = fetchJson;
 function poll(func, options) {